src/read_write_tradeoff/requirements.txt
python-dotenv
psycopg2-binary
Faker
azure-identity
azure-storage-blob
nltk
pandas